removed the minNotational, and minQuantity filter.
It makes absolutely no sense, but I had to try it. I disabled the
filters and ran it and suddenly it worked absolutely fine. This leads me
to have no idea why I was getting those errors in the past and why I am
no longer getting the errors now. But it does mean that I have solved a
known problem. This problem is that you can't really justify a buy on 87
other items by selling just one if you have to appease the minNotational
filter. Which meant that my algorithm would run, sell a few things, then
give up a few times. Now this is no longer an issue. Simply put, every
transaction is just working and I have no idea why.